Update to account fetching, web3 injection
Added functionality to fetch account information synchronously prior to
web3 injection, allowing accounts to be visible to DApps from the moment
they load.

Fixed web3 injection so it is only injected into a single frame in a
tab.

Removed vmSubprovider since it seemed to be bugged, unknown if this will
cause problems with eth_estimateGasCost or if that call will be
forwarded to rpc provider automatically.